+++ README.md
@@ -1,0 +1,34 @@
+
+Test task:  My Own Crawler
+----------------------------------------------------
+### The simplest web crawler
+    https://github.com/AndreiKviatkouski/MyOwnCrawler.git
+     
+Using
+--------
+
+    git clone  https://github.com/AndreiKviatkouski/MyOwnCrawler.git
+     
+Building
+--------
+ 
+ To build this project you will need Maven 2. You can get it at:
+ 
+     http://maven.apache.org
+
+ Clean compilation products:
+ 
+     mvn clean
+     
+ Compile:
+ 
+     mvn compile
+     
+  ## Technological Stack
+     *   Java 11LS
+     *   Project JSOUP 
+     *   Project Lombok
+     *   Project Junit-Jupiter
+   
+     
+ ## Repo owner: Andrei Kviatkouski
