+++ src/main/java/by/AndreiKviatkouski/actions/SpiderAction.java
@@ -1,0 +1,148 @@
+package by.AndreiKviatkouski.actions;
+
+import by.AndreiKviatkouski.entyties.Element;
+import by.AndreiKviatkouski.service.ElementService;
+import by.AndreiKviatkouski.service.SpiderService;
+
+import java.io.FileWriter;
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.util.*;
+
+public class SpiderAction {
+
+    private final Set<String> pagesVisited = new HashSet<>();
+    private PrintWriter pw;
+    private final Map<String, Integer> statistics = new HashMap<>();// statistic count words
+    private final List<Element> elements = new ArrayList<>();
+    ElementService elementService = new ElementService();
+
+    /**
+     * Our main launching point for the Spider's functionality. Internally it creates spider legs
+     * that make an HTTP request and parse the response (the web page).
+     *
+     * @param url   - The starting point of the spider
+     * @param words - The words or string that you are searching for
+     */
+
+
+    public void searchRecursive(String url, List<String> words, int MAX_DEEP, int MAX_PAGES_TO_SEARCH) {
+
+        SpiderService leg = new SpiderService();
+
+        int deepStart = 0;
+
+        if (this.pagesVisited.size() < MAX_PAGES_TO_SEARCH && deepStart < MAX_DEEP && !this.pagesVisited.contains(url)) {// start conditions
+            leg.crawl(url);
+
+            List<String> links = leg.getLinks();// create list link
+
+            for (String word : words) {
+                int count = leg.countWorlds(word);
+                statistics.put(word, count);
+            }
+
+            this.pagesVisited.add(url);// added link
+
+            this.printResult(url, words, statistics); // print
+
+            Element newElement = elementService.create(url, words, statistics);
+            elements.add(newElement);//add element
+
+            for (String link : links) {
+                searchRecursive(link, words, deepStart + 1, MAX_PAGES_TO_SEARCH);//set deep
+            }
+        }
+    }
+
+    public void openFile(String file) {
+        try {
+            FileWriter fw = new FileWriter(file);
+            pw = new PrintWriter(fw);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public void closeFile() {
+        pw.close();
+    }
+
+    public void printResult(String url, List<String> words, Map<String, Integer> statistics) {
+        StringBuilder builder = new StringBuilder(url + " ");
+        for (String key : words) {
+            Integer value = statistics.get(key);
+            builder.append(value).append(" ");
+        }
+        builder.append("\n");
+        pw.write(builder.toString());
+    }
+
+
+
+    public void printSortElementTopTen() {
+        elementService.sortDown(elements);
+        if (elements.size() >= 10) {
+            StringBuilder builder = new StringBuilder();
+            for (int i = 0; i < 10; i++) {
+                Element element = elements.get(i);
+                builder.append(element.getUrl()).append(" ").
+                        append(element.getValue1()).append(" ").
+                        append(element.getValue2()).append(" ").
+                        append(element.getValue3()).append(" ").
+                        append(element.getValue4()).append(" ").
+                        append(element.getValue5()).append(" ");
+
+                builder.append("Numbers are").append(" ");
+                builder.append("\n");
+
+                builder.append("     ").
+                        append(element.getParam1()).
+                        append("-").
+                        append(element.getValue1()).
+                        append(" ").
+                        append("hits");
+                builder.append("\n");
+
+                builder.append("     ").
+                        append(element.getParam2()).
+                        append("-").
+                        append(element.getValue2()).
+                        append(" ").
+                        append("hits");
+                builder.append("\n");
+
+                builder.append("     ").
+                        append(element.getParam3()).
+                        append("-").append(element.getValue3()).
+                        append(" ").
+                        append("hits");
+                builder.append("\n");
+
+                builder.append("     ").
+                        append(element.getParam4()).
+                        append("-").
+                        append(element.getValue4()).
+                        append(" ").
+                        append("hits");
+                builder.append("\n");
+
+                builder.append("     ").
+                        append(element.getParam5()).
+                        append("-").
+                        append(element.getValue5()).
+                        append(" ").
+                        append("hits");
+                builder.append("\n");
+                builder.append("\n");
+                builder.append("\n");
+
+            }
+
+            builder.append("\n");
+            pw.write(builder.toString());
+        } else {
+            System.err.println("Elements size is < 10!!!");
+        }
+    }
+}
